feat: :sparkles: Created the WI related to view user profile
The work item (WI) associated with the user story “As a user, I want to be able to click on a user's profile to see their uploaded datasets” has been completed. A new route has been added that allows users to view the datasets uploaded by a specific person. In the dataset view, the “Uploaded by: X” label is now clickable. When a user clicks on a person’s name, they are taken to that user’s profile page, where all datasets uploaded by them are displayed.  The view_dataset.html template was updated accordingly: the previous placeholder link (#) has been replaced with the new route, enabling this navigation.